requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
email_validator==2.1.0
gunicorn==21.2.0
pytest==8.2.0
WTForms==3.2.1
Werkzeug==3.1.3
SQLAlchemy==2.0.44
